Rough draft of report model populating for frontend, disk, buje.

+++ CheckupExec/Models/AnalysisModels/UCForecastModel.cs
@@ -11,7 +11,7 @@ namespace CheckupExec.Models.AnalysisModels
     {
         public string StorageName { get; set; }
 
-        public double TotalCapacity { get; set; }
+        public long TotalCapacity { get; set; }
 
         public List<UsedCapacity> UsedCapacityInstances { get; set; }
